[MO] - [refactor] -> refactor clients
