+++ systemtest/src/test/java/io/strimzi/systemtest/SecurityST.java
@@ -18,7 +18,7 @@ import io.strimzi.systemtest.annotations.OpenShiftOnly;
 import io.strimzi.systemtest.utils.MetricsUtils;
 import io.strimzi.systemtest.utils.StUtils;
 import io.strimzi.test.TestUtils;
-import io.strimzi.test.k8s.KubeClusterException;
+import io.strimzi.test.k8s.exceptions.KubeClusterException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hamcrest.Matchers;
@@ -125,7 +125,7 @@ class SecurityST extends MessagingBaseST {
                                 commandForZookeeperPort2888).out();
                 checkZookeeperCertificates(outputForZookeeperPort2888);
             } catch (KubeClusterException e) {
-                if (e.result != null && e.result.exitStatus() == 104) {
+                if (e.result != null && e.result.returnCode() == 104) {
                     LOGGER.info("The connection for {} was forcibly closed because of new zookeeper leader", KafkaResources.zookeeperPodName(CLUSTER_NAME, podId));
                 } else {
                     throw new RuntimeException(e);
