+++ test/src/main/java/io/strimzi/test/k8s/cluster/Minikube.java
@@ -2,12 +2,16 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.test.k8s;
+package io.strimzi.test.k8s.cluster;
 
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.strimzi.test.executor.Exec;
+import io.strimzi.test.k8s.KubeClient;
+import io.strimzi.test.k8s.exceptions.KubeClusterException;
+import io.strimzi.test.k8s.cmdClient.KubeCmdClient;
+import io.strimzi.test.k8s.cmdClient.Kubectl;
 
-import static io.strimzi.test.k8s.Minishift.CONFIG;
+import static io.strimzi.test.k8s.cluster.Minishift.CONFIG;
 
 /**
  * A {@link KubeCluster} implementation for {@code minikube} and {@code minishift}.
@@ -24,20 +28,11 @@ public class Minikube implements KubeCluster {
     @Override
     public boolean isClusterUp() {
         try {
-            return Exec.exec(CMD, "status").exitStatus() == 0;
+            return Exec.exec(CMD, "status").exitStatus();
         } catch (KubeClusterException e) {
+            e.printStackTrace();;
             return false;
         }
-    }
-
-    @Override
-    public void clusterUp() {
-        Exec.exec(CMD, "start");
-    }
-
-    @Override
-    public void clusterDown() {
-        Exec.exec(CMD, "stop");
     }
 
     @Override
@@ -53,5 +48,4 @@ public class Minikube implements KubeCluster {
     public String toString() {
         return CMD;
     }
-
 }
