+++ test/src/main/java/io/strimzi/test/k8s/cluster/OpenShift.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright Strimzi authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+package io.strimzi.test.k8s.cluster;
+
+import io.fabric8.openshift.client.DefaultOpenShiftClient;
+import io.strimzi.test.TestUtils;
+import io.strimzi.test.executor.Exec;
+import io.strimzi.test.k8s.KubeClient;
+import io.strimzi.test.k8s.exceptions.KubeClusterException;
+import io.strimzi.test.k8s.cmdClient.KubeCmdClient;
+import io.strimzi.test.k8s.cmdClient.Oc;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+import static io.strimzi.test.k8s.cluster.Minishift.CONFIG;
+
+public class OpenShift implements KubeCluster {
+
+    private static final Logger LOGGER = LogManager.getLogger(OpenShift.class);
+    private static final String OC = "oc";
+
+    @Override
+    public boolean isAvailable() {
+        return Exec.isExecutableOnPath(OC);
+    }
+
+    @Override
+    public boolean isClusterUp() {
+        try {
+            return Exec.exec(OC, "status").exitStatus();
+        } catch (KubeClusterException e) {
+            e.printStackTrace();
+            return false;
+        }
+    }
+
+    @Override
+    public KubeCmdClient defaultCmdClient() {
+        return new Oc();
+    }
+
+    public String toString() {
+        return OC;
+    }
+}
