+++ test/src/main/java/io/strimzi/test/k8s/cmdClient/Oc.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.test.k8s;
+package io.strimzi.test.k8s.cmdClient;
 
 import io.strimzi.test.executor.Exec;
 import org.apache.logging.log4j.LogManager;
@@ -22,9 +22,7 @@ public class Oc extends BaseCmdKubeClient<Oc> {
 
     private static final String OC = "oc";
 
-    Oc() {
-
-    }
+    public Oc() { }
 
     private Oc(String futureNamespace) {
         namespace = futureNamespace;
@@ -73,10 +71,6 @@ public class Oc extends BaseCmdKubeClient<Oc> {
         return this;
     }
 
-    public Oc newApp(String template) {
-        return newApp(template, emptyMap());
-    }
-
     public Oc newApp(String template, Map<String, String> params) {
         List<String> cmd = namespacedCommand("new-app", template);
         for (Map.Entry<String, String> entry : params.entrySet()) {
@@ -85,11 +79,6 @@ public class Oc extends BaseCmdKubeClient<Oc> {
         }
 
         Exec.exec(cmd);
-        return this;
-    }
-
-    public Oc newProject(String name) {
-        Exec.exec(namespacedCommand("new-project", name));
         return this;
     }
