+++ test/src/main/java/io/strimzi/test/k8s/exceptions/NoClusterException.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.test.k8s;
+package io.strimzi.test.k8s.exceptions;
 
 public class NoClusterException extends Exception {
     public NoClusterException(String message) {
