+++ topic-operator/src/test/java/io/strimzi/operator/topic/TopicOperatorBaseIT.java
@@ -18,8 +18,8 @@ import io.strimzi.api.kafka.model.KafkaTopicBuilder;
 import io.strimzi.api.kafka.model.status.Condition;
 import io.strimzi.test.BaseITST;
 import io.strimzi.test.TestUtils;
-import io.strimzi.test.k8s.KubeCluster;
-import io.strimzi.test.k8s.NoClusterException;
+import io.strimzi.test.k8s.cluster.KubeCluster;
+import io.strimzi.test.k8s.exceptions.NoClusterException;
 import io.vertx.core.Vertx;
 import kafka.server.KafkaConfig$;
 import org.apache.kafka.clients.admin.AdminClient;
